Ok, on further review, this thing technically just does html, no PHP class, so moving back to includes.
No matter how you slice it, it's clumsy, but this is easier without
introducing functional changes to the code.
